+++ CHANGES.rst
@@ -16,9 +16,10 @@ Changelog
 **Fixed**
 
 - #775 Analyses on Analysis Requests are hyperlinked to their Worksheets
+- #769 Traceback when submitting duplicate when Duplicate Variation is not set
+- #771 Slow Searches in Listing Views
 - #774 When retracting an Analysis Requests its analyses are also retracted
 - #772 Improved UID check in API
-
 
 **Security**